moved computation to chart
- computation of the raw data from the database is now performed in the chart component instead of the saga. This enables the chart to only query the database when the branch changes
- fixed bug where the chart bounds went over the screen edge when screen size changes (for example when resizing the developer console in chrome)
- #135
