+++ ui/src/visualizations/code-ownership/sagas/index.js
@@ -3,7 +3,7 @@ import { select, throttle, fork, takeEvery } from 'redux-saga/effects';
 import _ from 'lodash';
 
 import { fetchFactory, timestampedActionFactory, mapSaga } from '../../../sagas/utils.js';
-import { getCommitDataForSha, getOwnershipForCommits, getPreviousFilenames } from './helper.js';
+import { getCommitDataForSha, getOwnershipForCommits, getFilenamesForBranch, getPreviousFilenames } from './helper.js';
 
 //define actions
 export const requestCodeOwnershipData = createAction('REQUEST_CODE_OWNERSHIP_DATA');
@@ -21,7 +21,6 @@ export default function* () {
   yield fork(watchRefreshRequests);
   yield fork(watchRefresh);
   yield fork(watchSetCurrentBranch);
-  yield fork(watchSetActiveFiles);
 }
 
 //throttle ensures that only one refresh action will be dispatched in an interval of 2000ms
@@ -38,26 +37,18 @@ function* watchSetCurrentBranch() {
   yield takeEvery('CO_SET_CURRENT_BRANCH', mapSaga(requestRefresh));
 }
 
-function* watchSetActiveFiles() {
-  yield takeEvery('CO_SET_ACTIVE_FILES', mapSaga(requestRefresh));
-}
-
 //fetchFactory returns a function that calls the specified function*()
 // and dispatches the specified actions (requestCodeOwnershipData etc.) at appropriate points
 export const fetchCodeOwnershipData = fetchFactory(
   function* () {
     const state = yield select();
-    const activeFiles = state.visualizations.codeOwnership.state.config.activeFiles;
     const currentBranch = state.visualizations.codeOwnership.state.config.currentBranch;
 
-    const result = { ownershipData: [] };
+    const result = { rawData: [], previousFilenames:{} };
 
     if (
       currentBranch === null ||
-      currentBranch === undefined ||
-      activeFiles === null ||
-      activeFiles === undefined ||
-      activeFiles.length === 0
+      currentBranch === undefined
     ) {
       return result;
     }
@@ -68,6 +59,8 @@ export const fetchCodeOwnershipData = fetchFactory(
         if (!latestBranchCommit) {
           throw new Error('Latest branch commit not found');
         }
+
+        const activeFiles = await getFilenamesForBranch(currentBranch.branch);
 
         //get previous filenames for all active files
         const previousFilenames = await getPreviousFilenames(activeFiles, currentBranch);
@@ -81,63 +74,9 @@ export const fetchCodeOwnershipData = fetchFactory(
         //sort by date
         relevantOwnershipData = relevantOwnershipData.sort((a, b) => new Date(a.date) - new Date(b.date));
 
-        //stores the current ownership distribution for each file
-        const fileCache = {};
+        result.rawData = relevantOwnershipData;
+        result.previousFilenames = previousFilenames;
 
-        //step through the commits sequentially, starting with the oldest one
-        for (const commit of relevantOwnershipData) {
-          const commitResult = { sha: commit.sha, date: commit.date, ownership: {} };
-
-          //update fileCache for each file this commit touches
-          for (const file of commit.files) {
-            //if the file was deleted in this commit, delete it from the filecache
-            if (file.action === 'deleted') {
-              delete fileCache[file.path];
-            } else {
-              //if the file was either added or modified, we add it to the filecache (if it is relevant)
-              //the file is relevant if it is either one of the currently active files
-              // or if it is a previous version of an active file.
-              let relevant = activeFiles.includes(file.path);
-
-              if (!relevant) {
-                //look at the previous filenames of all active files
-                for (const [fileName, previousNames] of Object.entries(previousFilenames)) {
-                  if (relevant) break;
-                  //for all previous filenames of the file we are currently looking at
-                  for (const name of previousNames) {
-                    //if this old filename is the one the current commit touches
-                    // (same path and committed at a time where the file had that path),
-                    // this file is relevant
-                    if (
-                      name.oldFilePath === file.path &&
-                      new Date(name.hasThisNameFrom) <= new Date(commit.date) &&
-                      new Date(commit.date) <= new Date(name.hasThisNameUntil)
-                    ) {
-                      relevant = true;
-                      break;
-                    }
-                  }
-                }
-              }
-
-              if (relevant) {
-                fileCache[file.path] = file.ownership;
-              }
-            }
-          }
-
-          //now filecache stores the current ownership for each file that exists at the time of the current commit
-          for (const [filePath, fileOwnershipData] of Object.entries(fileCache)) {
-            for (const ownershipOfStakeholder of fileOwnershipData) {
-              if (commitResult.ownership[ownershipOfStakeholder.stakeholder]) {
-                commitResult.ownership[ownershipOfStakeholder.stakeholder] += ownershipOfStakeholder.ownedLines;
-              } else {
-                commitResult.ownership[ownershipOfStakeholder.stakeholder] = ownershipOfStakeholder.ownedLines;
-              }
-            }
-          }
-          result.ownershipData.push(commitResult);
-        }
         return result;
       })
       .catch((e) => {
